// back/src/incidents/dto/create-incident.dto.ts
export class CreateIncidentDto {
  _id: string;               
  serviceId: string;     
  titulo: string;
  descripcion: string;
  severidad: 'Baja' | 'Media' | 'Alta' | 'Crítica';
  estado: 'Abierto' | 'En progreso' | 'Resuelto';
  cadena: string;
  restaurante: string;
  fechaInicio: string;  
}